Refactor consolePresentationController to be progressive and keep track of physical screen line length, not (logial) content lines. Change updateDownloadSources to display uri without query strings

+++ GoodOfflineGames/Controllers/Measurement/FormattedStringMeasurementController.cs
@@ -13,7 +13,7 @@ namespace Controllers.Measurement
 {
     public class FormattedStringMeasurementController : IMeasurementController<string>
     {
-        public long Measure(string data)
+        public int Measure(string data)
         {
             return data.Length - Regex.Matches(data, Separators.ColorFormatting).Count * Separators.ColorFormatting.Length;
         }
